% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicDataSummary.R
\name{BasicDataSummary}
\alias{BasicDataSummary}
\title{Get general characteristics of data.frame.}
\usage{
BasicDataSummary(data, target, large = 1e+10, feat.perc = FALSE,
  na.perc = FALSE, class.perc = FALSE, large.perc = FALSE)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
Data to summarize.}

\item{target}{[\code{character(1)}]\cr
Target column. If not available please insert as \code{NULL}.}

\item{large}{[\code{numeric(1)}]\cr
What is a large (absolute) value for a numeric column?}

\item{feat.perc}{[\code{logical(1)}]\cr
Portion of features in percent? Default is \code{FALSE}.}

\item{na.perc}{[\code{logical(1)}]\cr
Portion of NAs in percent? Default is \code{FALSE}.}

\item{class.perc}{[\code{logical(1)}]\cr
Portion of classes in percent? Default is \code{FALSE}.}

\item{large.perc}{[\code{logical(1)}]\cr
Portion of large values in percent? Default is \code{FALSE}.}
}
\value{
[named \code{list}], containing characteristics.
}
\description{
Get general characteristics of data.frame.
}
\examples{
  data("iris")
  BasicDataSummary(iris, target = "Species")
}
