% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBasicReport.R
\name{createBasicReport}
\alias{createBasicReport}
\title{Creates a Basic Report Object}
\usage{
createBasicReport(id, data, target, ...)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
ID of the Task Object. \code{\link{makeBasicReportTask}}}

\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables}

\item{target}{[\code{character(1)}]\cr
Target column. If not available please insert as \code{NULL}.}

\item{...}{For now has no use}
}
\description{
Directly creates a report object with one command
}
