% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCatSumReport.R
\name{createCatSumReport}
\alias{createCatSumReport}
\title{Creates a CatSum Report Object}
\usage{
createCatSumReport(id, data, target, ...)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
ID of the Task Object. \code{\link{makeCatSumTask}}}

\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables}

\item{target}{[\code{character(1)}]\cr
Target column. If not available please insert as \code{NULL}.}

\item{...}{Further arguments for makeCatSumTask}
}
\description{
Directly creates a report object with one command
}
