% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCorrReport.R
\name{createCorrReport}
\alias{createCorrReport}
\title{Creates a Correlation Report Object}
\usage{
createCorrReport(id, data, ...)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
ID of the Task Object. \code{\link{makeCorrTask}}}

\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables}

\item{...}{For now it does nothing}
}
\description{
Directly creates a report object with one command
}
