% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFAReport.R
\name{createFAReport}
\alias{createFAReport}
\title{Creates a FA Report Object}
\usage{
createFAReport(id, data, ...)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
ID of the Task Object. \code{\link{makeFATask}}}

\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables}

\item{...}{as of now it does nothing}
}
\description{
Directly creates a report object with one command
}
