% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNumSumReport.R
\name{createNumSumReport}
\alias{createNumSumReport}
\title{Creates a NumSum Report Object}
\usage{
createNumSumReport(id, data, target, ...)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
ID of the Task Object. \code{\link{makeNumSumTask}}}

\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables}

\item{target}{[\code{character(1)}]\cr
Target column. If not available please insert as \code{NULL}.}

\item{...}{Further arguments for makeNumSumTask}
}
\description{
Directly creates a report object with one command
}
