% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastReport.R
\name{fastReport}
\alias{fastReport}
\title{Creates a full report for OpenML data sets}
\usage{
fastReport(data, target = NULL, reports = c("Basic", "CatSum", "NumSum",
  "Corr", "Cluster", "PCA", "MDS", "FA"), m.par.vals = list())
}
\arguments{
\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables}

\item{target}{[\code{character(1)}]\cr
Target column. If not available please insert as \code{NULL}.}

\item{reports}{[\code{character()}]\cr
The report types used on the dataset. Default will use all
available reports if applicable to the data types provided by
the dataset.}

\item{m.par.vals}{[\code{list()}]\cr
A names list containing parameters which will be passed to
the corresponding report function. [WIP]}
}
\description{
Imports a dataset from the openML data base and
creates a rmd report file
}
\examples{
# Report for the iris dataset
\dontrun{
data("iris")
fastReport(data = iris, target = "Species")
}
}
