% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finishReport.R
\name{finishReport}
\alias{finishReport}
\title{Creates the main Report with childs}
\usage{
finishReport(..., sub.dir = "Data_Report", save.mode = TRUE)
}
\arguments{
\item{...}{\cr
Report objects}

\item{sub.dir}{[\code{path}]\cr
The path where the child rmd files will be created}

\item{save.mode}{[\code{logical(1)}]\cr
In Save mode its not possible to use an existing folder.
To ensure no data is lost, a new folder will be created (if possible).}
}
\value{
creates rmd Files, returns NULL
}
\description{
The function writes the MainReport rmd file and write the child rmd files and
organize them together in one report
}
\examples{
data("airquality")
basic.report.task = makeReportTask(id = "AirqualityTask", data = airquality, target = "Wind")
basic.report = makeBasicReport(basic.report.task, data = airquality)

data("Boston", package = "MASS")
num.sum.task = makeNumSumTask(id = "BostonTask", data = Boston, target = "medv")
num.sum = makeNumSum(num.sum.task)
num.sum.report = makeNumSumReport(num.sum)

finishReport(basic.report, num.sum.report)
}
