% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCorrPlot.R
\name{generateCorrPlot}
\alias{generateCorrPlot}
\title{Generates code for a Correlation Plot}
\usage{
generateCorrPlot(corr.object, obj.name)
}
\arguments{
\item{corr.object}{[\code{CorrTask}]\cr
A corrTask Object}

\item{obj.name}{[\code{character()}]
the name of object from where to get the data.
make(Corr)Report function generates a random string for this.}
}
\value{
R Code as string
}
\description{
generates code for a Correlation Plot
}
\examples{
\dontrun{
corr.task = makeCorrTask(id = "cars", data = cars)
corr.result = makeCorr(corr.task)
generateCorrPlot(corr.result, "cars")
}
}
