% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePCAPlot.R
\name{generatePCAPlot}
\alias{generatePCAPlot}
\title{Generates code for a PCA Plot}
\usage{
generatePCAPlot(pca.obj, obj.name)
}
\arguments{
\item{pca.obj}{[\code{PCATask}]\cr
A PCATask Object}

\item{obj.name}{[\code{character()}]
the name of object from where to get the data.
make(PCA)Report function generates a random string for this.}
}
\value{
R Code as string
}
\description{
generates code for a PCA Plot
}
