% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCatSum.R
\name{getCatSum}
\alias{getCatSum}
\title{Computes the components and plots for the Categorial Summary}
\usage{
getCatSum(data, features, target, geombar.args)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables.}

\item{features}{[\code{character(length(catergorical.features))}]\cr
A character vector with length of the number of numeric features in the dataset.
This will be computed automatically when calling this function.}

\item{target}{[\code{character(1)}]\cr}

\item{geombar.args}{[\code{list()}] \cr
Other arguments to be passed to \link[ggplot2]{geom_bar}.
 The target column}
}
\value{
[\code{list()}]
  A list containing the categorical summary and ggplot for each categorical column
}
\description{
getCatSum calculates a few characteristics like frequency tables, contingency tables for categorical data and
 in Addition to that, for each column, a ggplot object will be implemented.
}
