% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterAnalysis.R
\name{getClusterAnalysis}
\alias{getClusterAnalysis}
\title{Computes the components and plots for the Cluster Analysis}
\usage{
getClusterAnalysis(data, num.features, method, par.vals, random.seed,
  scale.num.data, cluster.cols)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables.}

\item{num.features}{[\code{character(length(numeric.features))}]\cr
A character vector with length of the number of\cr
numeric features in the dataset.\cr
This will be computed automatically when calling this function.}

\item{method}{[\code{character(1L)}]\cr
Method used for clustering}

\item{par.vals}{[\code{list()}]\cr
Other arguments to be passed to selected method}

\item{random.seed}{[\code{integer(1L)}]\cr
Random seed for clustering method}

\item{scale.num.data}{[\code{logical(1L)}]\cr
Logical whether or not so scale numeric data\cr
for cluster analysis}

\item{cluster.cols}{[\code{character()}]\cr
Selected columns in a named character vector to apply cluster analysis\cr}
}
\value{
[\code{list()}]
  A list containing the cluster analysis and plot-code
}
\description{
getClusterAnalysis computes the cluster analysis of numeric data
}
