% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataType.R
\name{getDataType}
\alias{getDataType}
\title{Get Data Types from a data.frame.}
\usage{
getDataType(data, target)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
Data to extract the types from}

\item{target}{[\code{character(1)}]\cr
Target column. If not available please insert as \code{NULL}.}
}
\value{
[named \code{list}], containing vectors according to types. Vectors contain column names of the data.frame
}
\description{
Get Data Types from a data.frame.
}
\examples{
data("iris")
getDataType(iris, target = NULL)
}
