% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDbscanEps.R
\name{getEps}
\alias{getEps}
\title{Computes a suitable eps value for DBScan}
\usage{
getEps(data)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
A Dataframe with different numeric variables.}
}
\value{
[\code{numeric(1)}]
  An eps value for dbscan
}
\description{
The criterion used is an elbow criterion for knn distancies.
 Since this is a subjective criterion the calculation is just heuristic
}
