% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDSAnalysis.R
\name{getMDSAnalysis}
\alias{getMDSAnalysis}
\title{Computes the multidimensional scaling and plots}
\usage{
getMDSAnalysis(dist, method, par.vals)
}
\arguments{
\item{dist}{[\code{dist}]\cr
A distance matrix}

\item{method}{[\code{character(1L)}]\cr
Method used for multidimensional scaling}

\item{par.vals}{[\code{list()}]\cr
Other arguments to be passed to selected method}
}
\value{
[\code{list()}]
  A list containing the MDS analysis and plot
}
\description{
getMDSAnalysis computes the MDS analysis of a distance matrix
}
