% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumSum.R
\name{getNumSum}
\alias{getNumSum}
\title{Computes the components and plots for the Numeric Summary}
\usage{
getNumSum(data, features, target)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables.}

\item{features}{[\code{character(length(numeric.features))}]\cr
A character vector with length of the number of numeric features in the dataset.
This will be computed automatically when calling this function.}

\item{target}{[\code{character(1)}]\cr
The target column}
}
\value{
[\code{list()}]
  A list containing the numeric summary and ggplot for each numeric column
}
\description{
getNumSum applies function over columns in order to extract location- and scale parameters.
 In Addition to that, for each column, a ggplot object will be implemented.
}
