% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBasicReport.R
\name{makeBasicReport}
\alias{makeBasicReport}
\title{Creates a BasicReport Object [WIP]}
\usage{
makeBasicReport(report.task, data)
}
\arguments{
\item{report.task}{[\code{ReportTask}]\cr
A Report Task object containing general information of the data analysis problem}

\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables}
}
\value{
CorrTask
}
\description{
A Basic Report object containing the report task, a basic data summary and missing value summary
}
\examples{
my.task = makeCorrTask(id = "test", data = cars)
# Extract Data
my.task$env$data
 data("airquality")
 my.report.task = makeReportTask(id = "test.report", data = airquality, target = "Wind")
 basic.report = makeBasicReport(my.report.task, data = airquality)

}
