% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBasicReportTask.R
\name{makeBasicReportTask}
\alias{makeBasicReportTask}
\title{Creates a ReportTask Object}
\usage{
makeBasicReportTask(id, data, target, show.NA.msg = TRUE, ...)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
ID of the Task Object}

\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables}

\item{target}{[\code{character(1)}]\cr
Defines the target column. f the dataset does not contain a target column please insert \code{NULL}}

\item{show.NA.msg}{[\code{logical(1)}]\cr
Logical whether to show missing values message\cr
Default is \code{TRUE}.}

\item{...}{For now has no use}
}
\value{
[\code{BasicReportTask}]
}
\description{
A Report Task encapsulates the Data with some additional information
}
\examples{
basic.report.task = makeBasicReportTask(id = "report.test", data = iris, target = "Species")
# Extract Data
basic.report.task$env$data
}
