% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCatSum.R
\name{makeCatSum}
\alias{makeCatSum}
\title{Calculates a Categorical Summary}
\usage{
makeCatSum(cat.sum.task)
}
\arguments{
\item{cat.sum.task}{[\code{CatSumTask}]\cr
A CatSumTask Object}
}
\value{
CatSumObj
}
\description{
makeCatSum calculates a categorical summary and wrapes an object around it
}
\examples{
 data("Arthritis", package = "vcd")
 cat.sum.task = makeCatSumTask(id = "Arthritis.Task", data = Arthritis, target = "Improved")
 #get the categorical summary task object
 cat.sum.result = makeCatSum(cat.sum.task)
}
