% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCatSumReport.R
\name{makeCatSumReport}
\alias{makeCatSumReport}
\title{Creates an Categorical Summary Report object}
\usage{
makeCatSumReport(cat.sum.obj, type = NULL)
}
\arguments{
\item{cat.sum.obj}{[\code{CatSumObj} Object]\cr
A object of the "CatSumObj" class}

\item{type}{[\code{character(1)} Object]\cr
The type of the Report to create. AS of now the default is \code{NULL}}
}
\value{
CatSumReport
}
\description{
Creates an Categorical Summary Report object
}
\examples{
 data("Arthritis", package = "vcd")
 cat.sum.task = makeCatSumTask(id = "Arthritis.Task", data = Arthritis, target = "Improved")
 cat.sum.result = makeCatSum(cat.sum.task)
 #create the numeric summary report
 cat.sum.report = makeCatSumReport(cat.sum.result)
}
