% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCatSumTask.R
\name{makeCatSumTask}
\alias{makeCatSumTask}
\title{Creates a CategoricalSummaryTask Object (for factor, ordered and logical columns)}
\usage{
makeCatSumTask(id, data, target)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
ID of the Task Object}

\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables}

\item{target}{[\code{character(1)}]\cr
Target column. If not available please insert as \code{NULL}.}
}
\value{
CatSumTask object
}
\description{
A Task encapsulates the Data with some additional information
}
\examples{
 data("Arthritis", package = "vcd")
 cat.sum.task = makeCatSumTask(id = "Arthritis.Task", data = Arthritis, target = "Improved")
 #get Data
 cat.sum.task$env$data
}
