% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClusterAnalysis.R
\name{makeClusterAnalysis}
\alias{makeClusterAnalysis}
\title{Calculates a Cluster Analysis}
\usage{
makeClusterAnalysis(cluster.task)
}
\arguments{
\item{cluster.task}{[\code{ClusterTask}]\cr
A ClusterTask Object}
}
\value{
ClusterAnalysis Object
}
\description{
makeClusterAnalysis calculates a cluster analysis and\cr
wrapes an object around it
}
\examples{
\dontrun{
 cluster.task = makeClusterTask(id = "iris", data = iris,
   method = "cluster.kmeans")
 cluster.analysis.result = makeClusterAnalysis(cluster.task)
}
}
