% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClusterAnalysisReport.R
\name{makeClusterAnalysisReport}
\alias{makeClusterAnalysisReport}
\title{Creates an Cluster Analysis Report object}
\usage{
makeClusterAnalysisReport(cluster.analysis.obj)
}
\arguments{
\item{cluster.analysis.obj}{[\code{ClusterAnalysisObj} Object]\cr
A object of the "ClusterAnalysisObj" class}
}
\value{
A ClusterAnalysisReport Object
}
\description{
Creates an Cluster Analysis Report object
}
\examples{
\dontrun{
 cluster.task = makeClusterTask(id = "iris", data = iris,
   method = "cluster.kmeans")
 cluster.analysis.result = makeClusterAnalysis(cluster.task)
 cluster.analysis.report = makeClusterAnalysisReport(cluster.analysis.result)
}
}
