% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClusterTask.R
\name{makeClusterTask}
\alias{makeClusterTask}
\title{Creates a ClusterTask Object}
\usage{
makeClusterTask(id, data, cluster.cols = NULL, method = "cluster.kmeans",
  random.seed = set.seed(Sys.time()), scale.num.data = TRUE,
  par.vals = list(), show.NA.msg = FALSE, ...)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
ID of the Task Object}

\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables}

\item{cluster.cols}{[\code{character()}]\cr
Named character vector to specify clusters. This only holds for\cr
not hierarchical cluster methods. Default \code{cluster.cols = NULL}\cr .
In default mode only datasets with maximal 10 numeric columns, a\cr
cluster analysis will be with the combinations: choose(5,2).\cr
If the amount of numeric columns is above 10, only the cluster for the PCA\cr
for the numeric columns will be calculated and out ouf (k,2) combinations\cr
randomly selected 10 combinations of tuples}

\item{method}{[\code{character(1)}]\cr
Defines the clustering method
Possible choices are: \cr
For Hierarchical Clustering:
\itemize{
\item{\code{cluster.h}} - for more information @seealso \link[stats]{hclust}
\item{\code{cluster.agnes}} - for more information @seealso \link[cluster]{agnes}
\item{\code{cluster.diana}} - for more information @seealso \link[cluster]{diana}
}
For Partitioning Clustering:
\itemize{
\item{\code{cluster.kmeans}} - for more information @seealso \link[stats]{kmeans}
\item{\code{cluster.kkmeans}} - for more information @seealso \link[kernlab]{kkmeans}
\item{\code{cluster.pam}} - for more information @seealso \link[cluster]{pam}
}
For Model-Based Clustering:
\itemize{
\item{\code{cluster.dbscan}} - for more information @seealso \link[dbscan]{dbscan}
\item{\code{cluster.mod}} - for more information @seealso \link[mclust]{mclust}
}
Default is \code{method = "cluster.kmeans"}}

\item{random.seed}{[\code{integer(1)}]\cr
Default is \code{random.seed = set.seed(Sys.time())}}

\item{scale.num.data}{[\code{logical(1L)}]\cr
Logical whether to scale numeric data or not.\cr
Default is \code{scale= TRUE}}

\item{par.vals}{[\code{list}]\cr
Additional arguments handled over to cluster algorithm \code{method}.\cr
Default is empty list \code{par.vals = list()}}

\item{show.NA.msg}{[\code{logical(1)}]\cr
Logical whether to show missing values message\cr
Default is \code{FALSE}.}

\item{...}{For now has no use}
}
\value{
ClusterTask Object
}
\description{
A Task encapsulates the Data with some additional information.\cr
As of now clustering will be only made for numerical data. Categorical data is [WIP]
}
\examples{
cluster.task = makeClusterTask(id = "iris", data = iris,
  method = "cluster.kmeans",
  random.seed = 89L, par.vals = list(iter.max = 15L))
cluster.task2 = makeClusterTask(id = "iris", data = iris,
  method = "cluster.kmeans", random.seed = 89L,
  cluster.cols = c("Sepal.Length" = "Petal.Length",
  "Sepal.Width" = "Petal.Width"))
}
