% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCorr.R
\name{makeCorr}
\alias{makeCorr}
\title{Calculates the Correlation Matrix}
\usage{
makeCorr(corr.task)
}
\arguments{
\item{corr.task}{[\code{CorrTask}]\cr
A corrTask Object}
}
\value{
CorrObject
}
\description{
makeCorr claculates the corr matrix with the predefines methode and wrapes an object around it
}
\examples{
my.task = makeCorrTask(id = "test", data = cars)
my.corr = makeCorr(my.task)
my.corr$corr.matrix
}
