% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCorrReport.R
\name{makeCorrReport}
\alias{makeCorrReport}
\title{Creates an Report object}
\usage{
makeCorrReport(corr.obj)
}
\arguments{
\item{corr.obj}{[\code{CorrObj} Object]\cr
A object of the "CorrObj" class}
}
\value{
CorrReport
}
\description{
Creates an Report object
}
\examples{
corr.task = makeCorrTask(id = "test", data = cars)
corr.result = makeCorr(corr.task)
corr.report = makeCorrReport(corr.result)
}
