% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFA.R
\name{makeFA}
\alias{makeFA}
\title{Calculates a Factor Analysis}
\usage{
makeFA(fa.task)
}
\arguments{
\item{fa.task}{[\code{FATask}]\cr
A FATask Object}
}
\value{
FA Object
}
\description{
makeFA calculates a exploratory factor analysis and wrapes an object around it
}
\examples{
 library(psych)
 data(bfi)
 #take small sample of size 200L:
 bfi_small = bfi[sample(seq_len(nrow(bfi)), size = 200L), ]
 FA.task = makeFATask(id = "bfi", data = bfi_small)
 FA.result = makeFA(FA.task)
}
