% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFAReport.R
\name{makeFAReport}
\alias{makeFAReport}
\title{Creates an Factor Analysis Report Object}
\usage{
makeFAReport(fa.obj)
}
\arguments{
\item{fa.obj}{[\code{FAObj} Object]\cr
A object of the "FAObj" class}
}
\value{
A FAReport Object
}
\description{
Creates an Factor Analysis Report Object
}
\examples{
 library(psych)
 data(bfi)
 #take small sample of size 200L:
 bfi_small = bfi[sample(seq_len(nrow(bfi)), size = 200L), ]
 FA.task = makeFATask(id = "bfi", data = bfi_small)
 FA.result = makeFA(FA.task)
 FA.report = makeFAReport(FA.result)
}
