% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMDSAnalysis.R
\name{makeMDSAnalysis}
\alias{makeMDSAnalysis}
\title{Calculates a Multidimensional Scaling}
\usage{
makeMDSAnalysis(mds.task)
}
\arguments{
\item{mds.task}{[\code{MDSTask}]\cr
A MDSTask Object}
}
\value{
MDS Object
}
\description{
makeMDS calculates a multidimensional scaling and\cr
wrapes an object around it
}
\examples{
mds.task = makeMDSTask(id = "swiss", data = swiss)
mds.analysis.result = makeMDSAnalysis(mds.task)
}
