% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMDSAnalysisReport.R
\name{makeMDSAnalysisReport}
\alias{makeMDSAnalysisReport}
\title{Creates an Multidimensional Scaling Report Object}
\usage{
makeMDSAnalysisReport(mds.analysis.obj)
}
\arguments{
\item{mds.analysis.obj}{[\code{MDSAnalysisObj} Object]\cr
A object of the "MDSAnalysisObj" class}
}
\value{
A MDSAnalysisReport Object
}
\description{
Creates an Multidimensional Scaling Report Object
}
\examples{
 mds.task = makeMDSTask(id = "swiss", data = swiss)
 mds.analysis.result = makeMDSAnalysis(mds.task)
 mds.analysis.report = makeReport(mds.analysis.result)
}
