% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMDSTask.R
\name{makeMDSTask}
\alias{makeMDSTask}
\title{Creates a Multi Dimensional Scaling Task Object}
\usage{
makeMDSTask(id, data, dist.norm = "euclidean", method = "cmdscale",
  par.vals = list(), show.NA.msg = FALSE, ...)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
ID of the Task Object}

\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables}

\item{dist.norm}{[\code{character(1)}]\cr
Character indicating how the distance matrix will be computed.\cr
Possible values are:"euclidean", "maximum", "manhattan", \cr
"canberra", "binary" or "minkowski".}

\item{method}{[\code{character(1)}]\cr
Defines the MDS method
Possible choices are: \cr
For Metric Scaling:
\itemize{
\item{\code{cmdscale}} - for more information @seealso \link[stats]{cmdscale}
\item{\code{wcmdscale}} - for more information @seealso \link[vegan]{wcmdscale}
\item{\code{smacofSym}} - for more information @seealso \link[smacof]{smacofSym}
}
For Non-Metric Scaling:
\itemize{
\item{\code{isoMDS}} - for more information @seealso \link[MASS]{isoMDS}
\item{\code{sammon}} - for more information @seealso \link[MASS]{sammon}
}
Default is \code{method = "cmdscale"}}

\item{par.vals}{[\code{list}]\cr
Additional arguments handled over to MDS algorithm \code{method}.\cr
Default is empty list \code{par.vals = list()}}

\item{show.NA.msg}{[\code{logical(1)}]\cr
Logical whether to show missing values message\cr
Default is \code{FALSE}.}

\item{...}{For now has no use}
}
\value{
MDSTask Object
}
\description{
A Task encapsulates the Data with some additional information.\cr
As of now MDS performs classical (weighted) MDS and non-metric methods
}
\examples{
data(swiss)
mds.task = makeMDSTask(id = "swiss", data = swiss,
 dist.norm = "euclidean", method = "cmdscale", show.NA.msg = TRUE)
}
