% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNumSum.R
\name{makeNumSum}
\alias{makeNumSum}
\title{Calculates a Numeric Summary}
\usage{
makeNumSum(num.sum.task)
}
\arguments{
\item{num.sum.task}{[\code{NumTask}]\cr
A NumSumTask Object}
}
\value{
NumSumObj
}
\description{
makeNumSum calculates a numeric summary and wrapes an object around it
}
\examples{
 data("Boston", package = "MASS")
 num.sum.task = makeNumSumTask(id = "BostonTask", data = Boston, target = "medv")
 #get the numeric summary task object
 num.sum = makeNumSum(num.sum.task)
}
