% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNumSumReport.R
\name{makeNumSumReport}
\alias{makeNumSumReport}
\title{Creates an Numeric Summary Report object}
\usage{
makeNumSumReport(num.sum.obj, type = NULL)
}
\arguments{
\item{num.sum.obj}{[\code{NumSumObj} Object]\cr
A object of the "NumSumObj" class}

\item{type}{[\code{character(1)} Object]\cr
The type of the Report to create. AS of now the default is \code{NULL}}
}
\value{
NumSumReport
}
\description{
Creates an Numeric Summary Report object
}
\examples{
 data("Aids2", package = "MASS")
 num.sum.task = makeNumSumTask(id = "AidsTask", data = Aids2,
   target = "sex")
 #get the numeric summary task object
 num.sum.result = makeNumSum(num.sum.task)
 #create the numeric summary report
 num.sum.report = makeNumSumReport(num.sum.result)
}
