% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNumSumTask.R
\name{makeNumSumTask}
\alias{makeNumSumTask}
\title{Creates a NumericSummaryTask Object (for numeric and integer columns)}
\usage{
makeNumSumTask(id, data, target, geom.hist.args = list(bins = 30, alpha =
  0.4), geom.dens.args = list(size = 2, alpha = 0.4),
  geom.box.args = list(), show.NA.msg = FALSE, ...)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
ID of the Task Object}

\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables}

\item{target}{[\code{character(1)}]\cr
Target column. If not available please insert as \code{NULL}.}

\item{geom.hist.args}{[\code{list()}] \cr
Other arguments to be passed to \link[ggplot2]{geom_histogram}.
Default is \code{list(bins = 30, alpha = 0.4)}}

\item{geom.dens.args}{[\code{list()}] \cr
Other arguments to be passed to \link[ggplot2]{geom_density}.
Default is \code{list(size = 2, alpha = 0.4)}}

\item{geom.box.args}{[\code{list()}] \cr
Other arguments to be passed to \link[ggplot2]{geom_boxplot}.
Default is \code{NULL}. Default args from \link[ggplot2]{geom_boxplot} will be passed.}

\item{show.NA.msg}{[\code{logical(1)}]\cr
Logical whether to show missing values message\cr
Default is \code{FALSE}.}

\item{...}{For now has no use}
}
\value{
NumSumTask object
}
\description{
A Task encapsulates the Data with some additional information
}
\examples{
 data("Aids2", package = "MASS")
 num.sum.task = makeNumSumTask(id = "AidsTask", data = Aids2,
   target = "sex")
 #get Data
 num.sum.task$env$data
}
