% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePCATask.R
\name{makePCATask}
\alias{makePCATask}
\title{Creates a PCATask Object}
\usage{
makePCATask(id, data, target, vars = NULL, exclude = character(0),
  show.NA.msg = FALSE, ...)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
ID of the Task Object}

\item{data}{[\code{data.frame}]\cr
Data for PCA. Only numeric columns will be used and the target column excluded.}

\item{target}{[\code{character(1)}]\cr
Target column. If not available please insert as \code{NULL}.}

\item{vars}{[\code{character(1)}]\cr
Column names}

\item{exclude}{[\code{character}]
Names of inputs, which should be excluded. Default is none.}

\item{show.NA.msg}{[\code{logical(1)}]\cr
Logical whether to show missing values message\cr
Default is \code{FALSE}.}

\item{...}{Further arguments passed to \code{\link[stats]{prcomp}}}
}
\value{
PCATask
}
\description{
Principal Components Analysis (PCA). A Task encapsulates the Data with some additional information
}
\examples{
data("iris")
pca.task = makePCATask(id = "iris.try", data = iris, target = "Species",
                        tol = 1e-1, center = TRUE)
# get Data
pca.task$env$data
}
