% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_makeBasicReport.R
\name{makeReport.BasicReportTask}
\alias{makeReport.BasicReportTask}
\title{Creates a BasicReport Object [WIP]}
\usage{
\method{makeReport}{BasicReportTask}(analysis.obj)
}
\arguments{
\item{analysis.obj}{[\code{ReportTask}]\cr
A Report Task object containing general information of the data analysis problem}
}
\value{
[\code{BasicReport}]
}
\description{
A Basic Report Object will be calculated containing the report task, a basic data summary and missing value summary
}
\examples{
 data("airquality")
 basic.report.task = makeBasicReportTask(id = "test.report", data = airquality, target = "Wind")
 basic.report.result = makeReport(basic.report.task)

}
