% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_makeCatSumReport.R
\name{makeReport.CatSumObj}
\alias{makeReport.CatSumObj}
\title{Creates an Categorical Summary Report object}
\usage{
\method{makeReport}{CatSumObj}(analysis.obj)
}
\arguments{
\item{analysis.obj}{[\code{CatSumObj} Object]\cr
A object of the "CatSumObj" class}
}
\value{
CatSumReport
}
\description{
Creates an Categorical Summary Report object
}
\examples{
 data("Arthritis", package = "vcd")
 cat.sum.task = makeCatSumTask(id = "Arthritis.Task", data = Arthritis, target = "Improved")
 cat.sum.result = makeCatSum(cat.sum.task)
 #create the numeric summary report
 cat.sum.report = makeReport(cat.sum.result)
}
