% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_makeCorrReport.R
\name{makeReport.CorrObj}
\alias{makeReport.CorrObj}
\title{Creates an Correlation Report object}
\usage{
\method{makeReport}{CorrObj}(analysis.obj)
}
\arguments{
\item{analysis.obj}{[\code{CorrObj} Object]\cr
A object of the "CorrObj" class}
}
\value{
CorrReport
}
\description{
Creates an Correlation Report object
}
\examples{
corr.task = makeCorrTask(id = "test", data = cars)
corr.result = makeCorr(corr.task)
corr.report = makeReport(corr.result)
}
