% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_makeFAReport.R
\name{makeReport.FAObj}
\alias{makeReport.FAObj}
\title{Creates an Factor Analysis Report Object}
\usage{
\method{makeReport}{FAObj}(analysis.obj)
}
\arguments{
\item{analysis.obj}{[\code{FAObj} Object]\cr
A object of the "FAObj" class}
}
\value{
A FAReport Object
}
\description{
Creates an Factor Analysis Report Object
}
\examples{
 library(psych)
 data(bfi)
 #take small sample of size 200L:
 bfi_small = bfi[sample(seq_len(nrow(bfi)), size = 200L), ]
 FA.task = makeFATask(id = "bfi", data = bfi_small)
 FA.result = makeFA(FA.task)
 FA.report = makeReport(FA.result)
}
