% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_makeMDSAnalysisReport.R
\name{makeReport.MDSAnalysisObj}
\alias{makeReport.MDSAnalysisObj}
\title{Creates an Multidimensional Scaling Report Object}
\usage{
\method{makeReport}{MDSAnalysisObj}(analysis.obj)
}
\arguments{
\item{analysis.obj}{[\code{MDSAnalysisObj} Object]\cr
A object of the "MDSAnalyisObj" class}
}
\value{
A MDSAnalysisReport Object
}
\description{
Creates an Multidimensional Scaling Report Object
}
\examples{
 mds.task = makeMDSTask(id = "swiss", data = swiss)
 mds.analysis.result = makeMDSAnalysis(mds.task)
 mds.analysis.report = makeReport(mds.analysis.result)
}
