% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_makeNumSumReport.R
\name{makeReport.NumSumObj}
\alias{makeReport.NumSumObj}
\title{Creates an Numeric Summary Report object}
\usage{
\method{makeReport}{NumSumObj}(analysis.obj)
}
\arguments{
\item{analysis.obj}{[\code{NumSumObj} Object]\cr
A object of the "NumSumObj" class}
}
\value{
NumSumReport
}
\description{
Creates an Numeric Summary Report object
}
\examples{
 data("Aids2", package = "MASS")
 num.sum.task = makeNumSumTask(id = "AidsTask", data = Aids2,
   target = "sex")
 #get the numeric summary task object
 num.sum.result = makeNumSum(num.sum.task)
 #create the numeric summary report
 num.sum.report = makeReport(num.sum.result)
}
