% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_makePCAReport.R
\name{makeReport.PCAObj}
\alias{makeReport.PCAObj}
\title{Creates an PCA Report object}
\usage{
\method{makeReport}{PCAObj}(analysis.obj)
}
\arguments{
\item{analysis.obj}{[\code{PCAObj} Object]\cr
A object of the "PCAObj" class}
}
\value{
PCAReport
}
\description{
Creates an PCA Report object
}
\examples{
pca.task = makePCATask(id = "iris.test", data = iris, center = TRUE, target = "Species")
pca.result = makePCA(pca.task)
pca.report = makeReport(pca.result)
}
