% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReport.R
\name{makeReport}
\alias{makeReport}
\title{Creates a report object}
\usage{
makeReport(analysis.obj)
}
\arguments{
\item{analysis.obj}{[\code{Report} Object]\cr
An Analysis Object or in case of basic summary it is a basic.report.task}
}
\value{
a report.analysis.obj
}
\description{
Creates a report object
}
\examples{
  data("airquality")
  basic.report.task = makeBasicReportTask(id = "test.report", data = airquality, target = "Wind")
  basic.report.result = makeReport(basic.report.task)

  corr.report.task = makeCorrTask(id = "corr.report", data = airquality)
  corr.report.result = makeCorr(corr.report.task)
  corr.report = makeReport(corr.report.result)
}
