% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReportTask.R
\name{makeReportTask}
\alias{makeReportTask}
\title{Creates a ReportTask Object}
\usage{
makeReportTask(id, data, target)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
ID of the Task Object}

\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables}

\item{target}{[\code{character(1)}]\cr
Defines the target column. f the dataset does not contain a target column please insert \code{NULL}}
}
\value{
[\code{ReportTask}]
}
\description{
A Report Task encapsulates the Data with some additional information
}
\examples{
my.report.task = makeReportTask(id = "report.test", data = iris, target = "Species")
# Extract Data
my.report.task$env$data
}
