% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot for ggplot}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{If ggplot objects are not handed over as a list, insert them manually}

\item{plotlist}{[\code{list()}]\cr
A list of ggplot objects. Default is \code{plotlist = NULL}}

\item{cols}{[\code{integer(1)}]\cr
Number of columns for the grid}

\item{layout}{[\code{matrix()}]\cr
A matrix specifying the layout. If present, \code{cols} is ignored.}
}
\description{
Arranges ggplot objects into a grid
}
