% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplotPages.R
\name{multiplotPages}
\alias{multiplotPages}
\title{Wrapper for multiplot}
\usage{
multiplotPages(plotlist, k, no.cols, ...)
}
\arguments{
\item{plotlist}{[\code{list()}]\cr
A list of ggplot objects. Default is \code{plotlist = NULL}}

\item{k}{[\code{matrix()}]\cr
Maximum number of plots on one page}

\item{no.cols}{[\code{integer(1)}]\cr
Maximum number of columns}

\item{...}{Further arguments are handed over to multiplot}
}
\description{
Wrapper for multiplot with further plotlist splitting
}
