% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatDistr.R
\name{plotFeatDistr}
\alias{plotFeatDistr}
\title{Plots univariate distribution of a feature.}
\usage{
plotFeatDistr(data, target, col, geom.hist.args = list(bins = 30, alpha = 0.4,
  colour = "black"), geom.dens.args = list(size = 2, alpha = 0.4))
}
\arguments{
\item{data}{[\code{data.frame}]\cr
Data.}

\item{target}{[\code{character(1)}]\cr
Target column. If no target is available in the dataset please insert \code{NULL}}

\item{col}{[\code{character(1)} | \code{integer(1)}]\cr
Selected feature from \code{data}.}

\item{geom.hist.args}{[\code{list()}] \cr
Other arguments to be passed to \link[ggplot2]{geom_histogram}.}

\item{geom.dens.args}{[\code{list()}] \cr
Other arguments to be passed to \link[ggplot2]{geom_density}.}
}
\value{
A ggplot2 object. Print it to plot it.
}
\description{
Plots univariate distribution of a feature.
}
\examples{
 data("diamonds", package = "ggplot2")
 plotFeatDistr(diamonds, target = "cut", col = "carat")
}
