% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHist.R
\name{plotHist}
\alias{plotHist}
\title{Creates a histogram plot for one/ all numerical features(s) of a dataset.}
\usage{
plotHist(data, target, col = NULL, show.plot = FALSE, absolute = TRUE,
  alpha = 0.4, colour = "black", bins = 30L, ...)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
Data.}

\item{target}{[\code{character(1)}]\cr
Target column. If no target is available in the dataset please insert \code{NULL}}

\item{col}{[\code{character(1)} | \code{integer(1)}]\cr
Selected feature from \code{data}. If all numerical features should be printed insert \code{NULL}.
Default value is \code{NULL}}

\item{show.plot}{[\code{logical(1)}]
Logical whether the ggplot(s) should be displayed or not when executing this function.
Default is \code{FALSE}}

\item{absolute}{[\code{logical(1)}]
Whether the histogramm should plot absolute or relative frequencies. Default is \code{TRUE}}

\item{alpha}{[code{numeric(1)}]
Default is \code{alpha = 0.4}}

\item{colour}{[code{character(1)}]
Default is  \code{colour = "black"}}

\item{bins}{[code{integer(1)}]
Default is  \code{bins = 30L}}

\item{\dots}{other arguments to be passed to \link[ggplot2]{geom_histogram}.}
}
\value{
A ggplot2 object. Print it to plot it. [WIP if col is null]
}
\description{
Creates a histogramm plot of a numerical feature or all numerical features of a dataset .
}
\examples{
 data("Boston", package = "MASS")
 gghistplot = plotHist(Boston, target = "medv", col = "age")
 gghistAllplot = plotHist(Boston, target = "medv", col = NULL)
}
