% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBasicReport.R
\name{writeBasicReport}
\alias{writeBasicReport}
\title{Writes a rmd file for the Basic Report [WIP]}
\usage{
writeBasicReport(basic.report, sub.dir = "Data_Report", save.mode = TRUE)
}
\arguments{
\item{basic.report}{[\code{BasicReport} Object]\cr
The report Object which should be converted to an rmd file}

\item{sub.dir}{[\code{character(1)}]\cr
the name of the (relative) sub-directory where the data report will be saved. Default is \code{Data_Report}}

\item{save.mode}{[\code{logical(1)}]\cr
In Save mode its not possible to use an existing folder.
To ensure no data is lost, a new folder will be created (if possible).
Default is \code{TRUE}}
}
\value{
Invisible NULL
}
\description{
Writes a rmd file for the Basic Report [WIP]
}
\examples{
  data("airquality")
  my.report.task = makeReportTask(id = "test.report", data = airquality, target = "Wind")
  basic.report = makeBasicReport(my.report.task, data = airquality)
  writeBasicReport(basic.report)

  data("iris")
  my.report.task2 = makeReportTask(id = "test.report2", data = iris, target = "Species")
  basic.report2 = makeBasicReport(my.report.task2, data = iris)
  writeBasicReport(basic.report2)
}
