% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeNumSumReport.R
\name{writeNumSumReport}
\alias{writeNumSumReport}
\title{Writes a rmd file for the Numeric Summary Report [WIP]}
\usage{
writeNumSumReport(num.sum.report, sub.dir = "Data_Report", save.mode = TRUE)
}
\arguments{
\item{num.sum.report}{[\code{NumSumReport} Object]\cr
The report Object which should be converted to an rmd file}

\item{sub.dir}{[\code{character(1)}]\cr
the name of the (relative) sub-directory where the data report will be saved. Default is \code{Data_Report}}

\item{save.mode}{[\code{logical(1)}]\cr
In Save mode its not possible to use an existing folder.
To ensure no data is lost, a new folder will be created (if possible).
Default is \code{TRUE}}
}
\value{
Invisible NULL
}
\description{
Writes a rmd file for the Numeric Summary Report [WIP]
}
\examples{
#'  data("Boston", package = "MASS")
 num.sum.task = makeNumSumTask(id = "BostonTask", data = Boston, target = "medv")
 #get the numeric summary task object
 num.sum = makeNumSum(num.sum.task)
 #create the numeric summary report
 num.sum.report = makeNumSumReport(num.sum)
 #write the report
 writeNumSumReport(num.sum.report)
}
