% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCatSumReport.R
\name{writeReport.CatSumReport}
\alias{writeReport.CatSumReport}
\title{Writes a rmd file for the Categorical Summary Report [WIP]}
\usage{
\method{writeReport}{CatSumReport}(report, sub.dir = "Data_Report",
  save.mode = TRUE, override = FALSE)
}
\arguments{
\item{report}{[\code{CatSumReport} Object]\cr
The report Object which should be converted to an rmd file}

\item{sub.dir}{[\code{character(1)}]\cr
the name of the (relative) sub-directory where the data report will be saved. Default is \code{Data_Report}}

\item{save.mode}{[\code{logical(1)}]\cr
In Save mode its not possible to use an existing folder.
To ensure no data is lost, a new folder will be created (if possible).
Default is \code{TRUE}}

\item{override}{[\code{logical(1)}]\cr
override controls if the function is allowed to override
an existing rmd-file}
}
\value{
rmd-file location
}
\description{
Writes a rmd file for the Categorical Summary Report [WIP]
}
\examples{
 data("Arthritis", package = "vcd")
 cat.sum.task = makeCatSumTask(id = "Arthritis.Task", data = Arthritis, target = "Improved")
 cat.sum.result = makeCatSum(cat.sum.task)
 #create the numeric summary report
 cat.sum.report = makeCatSumReport(cat.sum.result)
 writeReport(cat.sum.report, save.mode = FALSE, override = TRUE)
}
