% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeClusterAnalysisReport.R
\name{writeReport.ClusterAnalysisReport}
\alias{writeReport.ClusterAnalysisReport}
\title{Writes a rmd file for the ClusterAnalysis Report [WIP]}
\usage{
\method{writeReport}{ClusterAnalysisReport}(report, sub.dir = "Data_Report",
  save.mode = TRUE, override = FALSE)
}
\arguments{
\item{report}{[\code{BasicReport} Object]\cr
The report Object which should be converted to an rmd file}

\item{sub.dir}{[\code{character(1)}]\cr
the name of the (relative) sub-directory where the data report will be saved.\cr
Default is \code{Data_Report}}

\item{save.mode}{[\code{logical(1)}]\cr
In Save mode its not possible to use an existing folder.\cr
To ensure no data is lost, a new folder will be created (if possible).
Default is \code{TRUE}}

\item{override}{[\code{logical(1)}]\cr
override controls if the function is allowed to override\cr
an existing rmd-file}
}
\value{
Invisible NULL
}
\description{
Writes a rmd file for the ClusterAnalysis Report [WIP]
}
\examples{
\dontrun{
  cluster.task = makeClusterTask(id = "iris", data = iris,
   method = "cluster.kmeans")
  cluster.analysis.result = makeClusterAnalysis(cluster.task)
  cluster.analysis.report = makeReport(cluster.analysis)
  writeReport(report, save.mode = FALSE, override = TRUE)
}
}
