% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFAReport.R
\name{writeReport.FAReport}
\alias{writeReport.FAReport}
\title{Writes a rmd file for the Factor Analysis Report [WIP]}
\usage{
\method{writeReport}{FAReport}(report, sub.dir = "Data_Report",
  save.mode = TRUE, override = FALSE)
}
\arguments{
\item{report}{[\code{FAReport} Object]\cr
The report Object which should be converted to an rmd file}

\item{sub.dir}{[\code{character(1)}]\cr
the name of the (relative) sub-directory where the data report will be saved.\cr
Default is \code{Data_Report}}

\item{save.mode}{[\code{logical(1)}]\cr
In Save mode its not possible to use an existing folder.
To ensure no data is lost, a new folder will be created (if possible).}

\item{override}{[\code{logical(1)}]\cr
override controls if the function is allowed to override
an existing rmd-file
Default is \code{TRUE}}
}
\value{
Invisible NULL
}
\description{
Writes a rmd file for the Factor Analysis Report [WIP]
}
\examples{
 library(psych)
 data(bfi)
 #take small sample of size 200L:
 bfi_small = bfi[sample(seq_len(nrow(bfi)), size = 200L), ]
 FA.task = makeFATask(id = "bfi", data = bfi_small)
 FA.result = makeFA(FA.task)
 #create the report
 FA.report = makeReport(FA.result)
 #write the report
 writeReport(FA.report, save.mode = FALSE, override = TRUE)
}
