% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeNumSumReport.R
\name{writeReport.NumSumReport}
\alias{writeReport.NumSumReport}
\title{Writes a rmd file for the Numeric Summary Report [WIP]}
\usage{
\method{writeReport}{NumSumReport}(report, sub.dir = "Data_Report",
  save.mode = TRUE, override = FALSE)
}
\arguments{
\item{report}{[\code{NumSumReport} Object]\cr
The report Object which should be converted to an rmd file}

\item{sub.dir}{[\code{character(1)}]\cr
the name of the (relative) sub-directory where the data report will be saved. Default is \code{Data_Report}}

\item{save.mode}{[\code{logical(1)}]\cr
In Save mode its not possible to use an existing folder.
To ensure no data is lost, a new folder will be created (if possible).}

\item{override}{[\code{logical(1)}]\cr
override controls if the function is allowed to override
an existing rmd-file
Default is \code{TRUE}}
}
\value{
Invisible NULL
}
\description{
Writes a rmd file for the Numeric Summary Report [WIP]
}
\examples{
 data("Boston", package = "MASS")
 num.sum.task = makeNumSumTask(id = "BostonTask", data = Boston, target = "medv")
 #get the numeric summary task object
 num.sum = makeNumSum(num.sum.task)
 #create the numeric summary report
 report = makeNumSumReport(num.sum)
 #write the report
 writeReport(report, save.mode = FALSE, override = TRUE)
}
