% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePCAReport.R
\name{writeReport.PCAReport}
\alias{writeReport.PCAReport}
\title{Writes a rmd file for the report [WIP]}
\usage{
\method{writeReport}{PCAReport}(report, sub.dir = "Data_Report",
  save.mode = TRUE, override = FALSE)
}
\arguments{
\item{report}{[\code{*Report} Object]\cr
The report Object which should be converted to an rmd file}

\item{sub.dir}{[\code{character(1)}]\cr
the name of the directory where the data report will be saved}

\item{save.mode}{[\code{logical(1)}]\cr
In Save mode its not possible to use an existing folder.
To ensure no data is lost, a new folder will be created (if possible).}

\item{override}{[\code{logical(1)}]\cr
override controls if the function is allowed to override
an existing rmd-file}
}
\value{
rmd-file location
}
\description{
Writes a rmd file for the report [WIP]
}
\examples{
  pca.task = makePCATask(id = "Probe", data = iris, target = "Species",
                        tol = 1e-1, center = TRUE)
  pca.result = makePCA(pca.task)
  pca.report = makeReport(pca.result)
  writeReport(pca.report, save.mode = FALSE, override = TRUE)
}
