% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeReport.R
\name{writeReport}
\alias{writeReport}
\title{Writes a rmd file Report [WIP]}
\usage{
writeReport(report, sub.dir, save.mode, override)
}
\arguments{
\item{report}{[\code{Report} Object]\cr
A report Object}

\item{sub.dir}{[\code{character(1)}]\cr
the name of the (relative) sub-directory where the data report will be saved. Default is \code{Data_Report}}

\item{save.mode}{[\code{logical(1)}]\cr
In Save mode its not possible to use an existing folder.
To ensure no data is lost, a new folder will be created (if possible).
Default is \code{TRUE}}

\item{override}{[\code{logical(1)}]\cr
override controls if the function is allowed to override
an existing rmd-file}

\item{...}{[\code{character(1)}]\cr
for now its not used}
}
\value{
Invisible NULL; creates a rmd report file
}
\description{
Writes a rmd file Report [WIP]
}
\examples{
  data("airquality")
  report.task = makeBasicReportTask(id = "test.report", data = airquality, target = "Wind")
  basic.report = makeReport(report.task)
  writeReport(basic.report, save.mode = FALSE, override = TRUE)

  corr.report.task = makeCorrTask(id = "corr.report", data = airquality, type = "CorrPlot")
  corr.report.result = makeCorr(corr.report.task)
  corr.report = makeReport(corr.report.result)
  writeReport(corr.report, save.mode = FALSE, override = TRUE)

  # Or put the reports together
  finishReport(basic.report, corr.report, save.mode = FALSE, override = TRUE)


}
