% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finishReport.R
\name{finishReport}
\alias{finishReport}
\title{Creates the main Report with childs}
\usage{
finishReport(..., sub.dir = "Data_Report", save.mode = TRUE,
  theme = "cosmo", df.print = "paged", override = FALSE)
}
\arguments{
\item{...}{\cr
Report objects}

\item{sub.dir}{[\code{path}]\cr
The path where the child rmd files will be created}

\item{save.mode}{[\code{logical(1)}]\cr
In Save mode its not possible to use an existing folder.
To ensure no data is lost, a new folder will be created (if possible).}

\item{theme}{[\code{character()}]\cr
This param is the theme of the YAML-header. If set to NULL no theme will
be used}

\item{df.print}{[\code{character()}]\cr
This param sets the YAML header for how Dataframed should be printed.
If set to NULL df.print param will not be used.}

\item{override}{[\code{logical(1)}]\cr
override controls if the function is allowed to override
an existing rmd-file}
}
\value{
creates rmd Files, returns NULL
}
\description{
The function writes the MainReport rmd file and write the child rmd files and
organize them together in one report
}
\examples{

data("airquality")
basic.report.task = makeBasicReportTask(id = "AirqualityTask", data = airquality, target = "Wind")
basic.report = makeReport(basic.report.task)

data("Boston", package = "MASS")
num.sum.task = makeNumSumTask(id = "BostonTask", data = Boston, target = NULL)
num.sum.result = makeNumSum(num.sum.task)
num.sum.report = makeReport(num.sum.result)

corr.task = makeCorrTask(id = "test", data = cars)
corr.result = makeCorr(corr.task)
corr.report = makeReport(corr.result)
data(diamonds, package = "ggplot2")
corr.task2 = makeCorrTask(id = "test2", data = diamonds)
corr.result2 = makeCorr(corr.task2)
corr.report2 = makeReport(corr.result2)

data("Arthritis", package = "vcd")
cat.sum.task = makeCatSumTask(id = "Arthritis.Task", data = Arthritis,
  target = "Improved", na.rm = TRUE)
cat.sum.result = makeCatSum(cat.sum.task)
cat.sum.report = makeReport(cat.sum.result)

#combine all reports
finishReport(basic.report, num.sum.report, corr.report, corr.report2,
  cat.sum.report, save.mode = FALSE, override = TRUE)

}
