% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumSum.R
\name{getNumSum}
\alias{getNumSum}
\title{Computes the components and plots for the Numeric Summary}
\usage{
getNumSum(data, features, target, geom.hist.args, geom.dens.args, geom.box.args)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables.}

\item{features}{[\code{character(length(numeric.features))}]\cr
A character vector with length of the number of numeric features in the dataset.
This will be computed automatically when calling this function.}

\item{target}{[\code{character(1)}]\cr
The target column}

\item{geom.hist.args}{[\code{list()}] \cr
Other arguments to be passed to \link[ggplot2]{geom_histogram}.
Inserted in \link[AEDA]{makeNumSumTask}}

\item{geom.dens.args}{[\code{list()}] \cr
Other arguments to be passed to \link[ggplot2]{geom_density}.
Inserted in \link[AEDA]{makeNumSumTask}}

\item{geom.box.args}{[\code{list()}] \cr
Other arguments to be passed to \link[ggplot2]{geom_boxplot}.
Inserted in \link[AEDA]{makeNumSumTask}}
}
\value{
[\code{list()}]
 A list containing the numeric summary and ggplot for each numeric column
}
\description{
getNumSum applies function over columns in order to extract location- and scale parameters.\cr
 \cr
 In Addition to that, for each column, a ggplot object will be implemented.
 For the numeric summary data.frame following columns will be explained:\cr

 The sample kurtosis will be calculated via \eqn{1/n * \sum_{i=1}^{n}(x_i - \bar{x} / s)^{4}}\cr
 The sample skewness will be calculated via \eqn{1/n * \sum_{i=1}^{n}(x_i - \bar{x}  /s)^{3}}\cr
 l.bound is defined as \eqn{q_{0.25}} - 1.5IQR and lbound is defined as \eqn{q_{0.75}} + 1.5IQR \cr
 where IQR = \eqn{q_{0.75} - q{0.25}}\cr
}
