% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBasicReport.R
\name{makeBasicReport}
\alias{makeBasicReport}
\title{Creates a BasicReport Object [WIP]}
\usage{
makeBasicReport(basic.report.task)
}
\arguments{
\item{basic.report.task}{[\code{BasicReportTask}]\cr
A Basic Report Task object containing general information of the data analysis problem}
}
\value{
[\code{BasicReport}]\cr
}
\description{
A Basic Report Object will be calculated containing the report task, a basic data summary and missing value summary
}
\examples{
 data("airquality")
 basic.report.task = makeBasicReportTask(id = "test.report", data = airquality, target = "Wind")
 basic.report.result = makeBasicReport(basic.report.task)

}
