% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCorr.R
\name{makeCorr}
\alias{makeCorr}
\title{Calculates the Correlation Matrix}
\usage{
makeCorr(corr.task)
}
\arguments{
\item{corr.task}{[\code{CorrTask}]\cr
A corrTask Object}
}
\value{
CorrObject
}
\description{
makeCorr calculates the correlation matrix with the predefines methode and wrapes an object around it
}
\examples{
corr.task = makeCorrTask(id = "test", data = cars)
corr.result = makeCorr(corr.task)
corr.result$corr.matrix
}
