% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCorrTask.R
\name{makeCorrTask}
\alias{makeCorrTask}
\title{Creates a CorrTask Objects}
\usage{
makeCorrTask(id, data, method = "pearson", vars = NULL, type = "CorrPlot")
}
\arguments{
\item{id}{[\code{character(1)}]\cr
ID of the Task Object}

\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables}

\item{method}{[\code{character(1)}]\cr
Defines the correlation method
Possible choices are:
\dQuote{pearson}}

\item{vars}{[\code{character(1)}]\cr
Column names to use for correlation}

\item{type}{[\code{character(1)}]\cr
The type of the Report to create. Example: "CorrPlot"}
}
\value{
CorrTask
}
\description{
A Task encapsulates the Data with some additional information
}
\examples{
my.task = makeCorrTask(id = "test", data = cars)
# Extract Data
my.task$env$data
}
