% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNumSumTask.R
\name{makeNumSumTask}
\alias{makeNumSumTask}
\title{Creates a NumericSummaryTask Object (for numeric and integer columns)}
\usage{
makeNumSumTask(id, data, target)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
ID of the Task Object}

\item{data}{[\code{data.frame}]\cr
A Dataframe with different variables}

\item{target}{[\code{character(1)}]\cr
Target column. If not available please insert as \code{NULL}.}
}
\value{
NumericSummaryTask
}
\description{
A Task encapsulates the Data with some additional information
}
\examples{
 data("Boston", package = "MASS")
 numTask = makeNumSumTask(id = "BostonTask", data = Boston, target = "medv")
 #get Data
 numTask$env$data
}
