% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePCA.R
\name{makePCA}
\alias{makePCA}
\title{Calculates the PCA}
\usage{
makePCA(pca.task)
}
\arguments{
\item{pca.task}{[\code{PCATask}]\cr
A pcaTask Object}
}
\value{
PCAObject
}
\description{
Principal components analysis on the given numeric data matrix and returns the results as an object
}
\examples{
my.task = makePCATask(id = "test", data = cars)
my.pca1 = makePCA(my.task)

iris<-iris[,-5] # 'x' must be numeric
test.task = makePCATask(id = "Probe", data = iris, target = "Petal.Lenght",
            scale = TRUE, tol = 1e-1)
my.pca2 = makePCA(test.task)
}
