% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePCA.R
\name{makePCA}
\alias{makePCA}
\title{Calculates the PCA}
\usage{
makePCA(pca.task)
}
\arguments{
\item{pca.task}{[\code{PCATask}]\cr
A pcaTask Object}
}
\value{
PCAObject
}
\description{
Principal components analysis on the given numeric data matrix and returns the results as an object
}
\examples{
pca.task = makePCATask(id = "iris.test", data = iris, target = "Species",
                        tol = 1e-1, center = TRUE)
pca.result = makePCA(pca.task)
}
