% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePCAReport.R
\name{makePCAReport}
\alias{makePCAReport}
\title{Creates an Report object}
\usage{
makePCAReport(pca.obj)
}
\arguments{
\item{pca.obj}{[\code{PCAObj} Object]\cr
A object of the "PCAObj" class}
}
\value{
PCAReport
}
\description{
Creates an Report object
}
\examples{
test.task = makePCATask(id = "Probe", data = iris, center = TRUE)
my.pca2 = makePCA(test.task)
report1 = makePCAReport(my.pca2)
}
