% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePCATask.R
\name{makePCATask}
\alias{makePCATask}
\title{Creates a PCATask Objects}
\usage{
makePCATask(id, data, target, method = "all", vars = NULL,
  exclude = character(0), ...)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
ID of the Task Object}

\item{data}{[\code{data.frame}]\cr
Data for PCA. Only numeric columns will be used and the target column excluded.}

\item{target}{[\code{character(1)}]\cr
Target column. If not available please insert as \code{NULL}.}

\item{method}{[\code{character(1)}]\cr
Defines the PCA method
\dQuote{all}}

\item{vars}{[\code{character(1)}]\cr
Column names}

\item{exclude}{[\code{character}]
Names of inputs, which should be excluded. Default is none.}

\item{...}{Further arguments passed to \code{\link[stats]{prcomp}}}
}
\value{
PCATask
}
\description{
Principal Components Analysis (PCA). A Task encapsulates the Data with some additional information
}
\examples{
data("iris")
test.task = makePCATask(id = "Probe", data = iris, target = "Petal.Length",
                        tol = 1e-1, center = TRUE)
# get Data
test.task$env$data
}
