% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot for ggplot}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{If ggplot objects are not handed over as a list, insert them manually}

\item{plotlist}{[\code{list()}]\cr
A list of ggplot objects. Default is \code{plotlist = NULL}}

\item{cols}{[\code{integer(1)}]\cr
Number of columns for the grid}

\item{layout}{[\code{matrix()}]\cr
A matrix specifying the layout. If present, \code{cols} is ignored.}
}
\description{
Arranges ggplot objects into a grid
}
\examples{
library(ggplot2)
p1 = ggplot(ChickWeight, aes(x = Time, y = weight, colour = Diet, group=Chick)) +
  geom_line() +
  ggtitle("Growth curve for individual chicks")

# Second plot
p2 = ggplot(ChickWeight, aes(x = Time, y = weight, colour = Diet)) +
  geom_point(alpha = 0.3) +
  geom_smooth(alpha = 0.2, size = 1) +
  ggtitle("Fitted growth curve per diet")

# Third plot
p3 = ggplot(subset(ChickWeight, Time == 21), aes(x = weight, colour = Diet)) +
  geom_density() +
  ggtitle("Final weight, by diet")

# Fourth plot
p4 = ggplot(subset(ChickWeight, Time == 21), aes(x = weight, fill = Diet)) +
 geom_histogram(colour = "black", binwidth = 50) +
 facet_grid(Diet ~ .) +
 ggtitle("Final weight, by diet") +
 theme(legend.position="none")
#multiplot the ggplots
multiplot(p1, p2, p3, p4, cols = 2)

}
