% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBar.R
\name{plotBar}
\alias{plotBar}
\title{Creates a bar plot of a categorical (discrete) feature.}
\usage{
plotBar(data, target, col = NULL, show.plot = FALSE)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
Data.}

\item{target}{[\code{character(1)}]\cr
Target column. If no target is available in the dataset please insert \code{NULL}}

\item{col}{[\code{character(1)} | \code{integer(1)}]\cr
Selected feature from \code{data}. If all categorical features should be printed insert \code{NULL}.
Default value is \code{NULL}}

\item{show.plot}{[\code{logical(1)}]
Logical whether the ggplot(s) should be displayed or not when executing this function.
Default is \code{FALSE}}

\item{\dots}{other arguments to be passed to \link[ggplot2]{geom_bar}.}
}
\value{
A ggplot2 object. Print it to plot it. [WIP if col is null]
}
\description{
Creates a bar plot of a categorical feature or all categorical features of a dataset .
}
\examples{
 data("diamonds", package = "ggplot2")
 ggplot = plotBar(diamonds, target = "cut", col = "carat")
}
