% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatDistr.R
\name{plotFeatDistr}
\alias{plotFeatDistr}
\title{Plots univariate distribution of a feature.}
\usage{
plotFeatDistr(data, target, col)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
Data.}

\item{target}{[\code{character(1)}]\cr
Target column. If no target is available in the dataset please insert \code{NULL}}

\item{col}{[\code{character(1)} | \code{integer(1)}]\cr
Selected feature from \code{data}.}
}
\value{
A ggplot2 object. Print it to plot it.
}
\description{
Plots univariate distribution of a feature.
}
