% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBasicReport.R
\name{writeReport.BasicReport}
\alias{writeReport.BasicReport}
\title{Writes a rmd file for the Basic Report [WIP]}
\usage{
\method{writeReport}{BasicReport}(report, sub.dir = "Data_Report",
  save.mode = TRUE, override = FALSE)
}
\arguments{
\item{report}{[\code{BasicReport} Object]\cr
The report Object which should be converted to an rmd file}

\item{sub.dir}{[\code{character(1)}]\cr
the name of the (relative) sub-directory where the data report will be saved. Default is \code{Data_Report}}

\item{save.mode}{[\code{logical(1)}]\cr
In Save mode its not possible to use an existing folder.
To ensure no data is lost, a new folder will be created (if possible).
Default is \code{TRUE}}

\item{override}{[\code{logical(1)}]\cr
override controls if the function is allowed to override
an existing rmd-file}
}
\value{
Invisible NULL
}
\description{
Writes a rmd file for the Basic Report [WIP]
}
\examples{
  data("airquality")
  my.report.task = makeBasicReportTask(id = "test.report", data = airquality, target = "Wind")
  report = makeReport(my.report.task)
  writeReport(report, save.mode = FALSE, override = TRUE)

  data("iris")
  my.report.task2 = makeReportTask(id = "test.report2", data = iris, target = "Species")
  report2 = makeReport(my.report.task2)
  writeReport(report2, save.mode = FALSE, override = TRUE)
}
